test(#5961): 회귀 시험을 tests/cases 로 옮긴다
`rust-unit-test-tiers --check` 가 PR base 대비 source-side 시험 총량 증가를 막는다
(4223 > 4221). 새 시험은 `tests/cases/` 가 지정된 자리이므로 `src/model/paragraph/
tests.rs` 에 넣었던 2건을 그리로 옮겼다.

`replace_line_segs()` 가 축 보정폭을 0 으로 되돌린다는 계약은 그 메서드가 crate
내부라 통합 시험에서 직접 부를 수 없다. 계약은 메서드 주석이 지키고, 어겼을 때의
증상(두 번 더해진 축)은 투영 시험들이 잡는다.

검증: 신규 회귀 6건 통과 · fmt · rust-test-suite-manifest · rust-unit-test-tiers
(4221, base 와 동일) 통과.
