+++ src/model/paragraph/tests.rs
@@ -1551,72 +1551,3 @@ fn shift_for_inline_control_insert_leaves_earlier_lines_alone() {
     assert_eq!(para.line_segs[1].text_start, 4);
     assert_eq!(para.line_segs[2].text_start, 16);
 }
-
-/// [#5961] 줄을 다시 계산하면 축 보정폭은 무효다.
-///
-/// `hwpx_axis_shift` 는 **파일에서 읽은** 줄이 HWP5 축보다 얼마나 짧은지를 적어 둔 값이다.
-/// reflow 가 만든 줄은 `char_offsets` 와 같은 HWP5 축에서 나오므로 보정할 것이 없고,
-/// 그대로 두면 다음 투영에서 이중으로 더해진다.
-#[test]
-fn replacing_line_segs_clears_the_hwpx_axis_shift() {
-    let mut para = Paragraph {
-        hwpx_axis_shift: 8,
-        line_segs: vec![
-            LineSeg {
-                text_start: 0,
-                ..Default::default()
-            },
-            LineSeg {
-                text_start: 24,
-                ..Default::default()
-            },
-        ],
-        ..Default::default()
-    };
-    assert_eq!(
-        para.line_seg_text_start(1),
-        32,
-        "선행 조건: 투영이 걸려 있다"
-    );
-
-    para.replace_line_segs(vec![
-        LineSeg {
-            text_start: 0,
-            ..Default::default()
-        },
-        LineSeg {
-            text_start: 32,
-            ..Default::default()
-        },
-    ]);
-
-    assert_eq!(
-        para.hwpx_axis_shift, 0,
-        "줄을 새로 계산했는데 보정폭이 남았다 — 다음 투영에서 이중으로 더해진다."
-    );
-    assert_eq!(
-        para.line_seg_text_start(1),
-        32,
-        "새 줄은 이미 HWP5 축이므로 투영이 값을 바꾸면 안 된다."
-    );
-}
-
-/// 보정폭이 0 인 문단(HWP5·HWP3·HML 출처)은 투영이 항등이다.
-#[test]
-fn a_zero_shift_projection_is_the_identity() {
-    let para = Paragraph {
-        line_segs: vec![
-            LineSeg {
-                text_start: 0,
-                ..Default::default()
-            },
-            LineSeg {
-                text_start: 48,
-                ..Default::default()
-            },
-        ],
-        ..Default::default()
-    };
-    assert_eq!(para.line_seg_text_start(0), 0);
-    assert_eq!(para.line_seg_text_start(1), 48);
-}
